refactor: rename InputReader > ConsoleInputManager
